chapter 10 - 10.12 exercise

+++ Code Samples/Chapter 10/favourite_number.py
@@ -2,6 +2,14 @@ from pathlib import Path
 import json
 
 path = Path('favourite_number.json')
-number = int(input('What is your favourite number? '))
-contents = json.dumps(number)
-path.write_text(contents)
+
+if path.exists():
+    contents = path.read_text()
+    number = json.loads(contents)
+    print(f'I know your favourite number. It\'s {number}!')
+
+else:
+    number = int(input('What is your favourite number? '))
+    contents = json.dumps(number)
+    path.write_text(contents)
+
